# Conflicts:
#	src/edit.py

+++ src/edit.py
@@ -1,11 +1,27 @@
-import string
 from flask import Flask, flash, render_template, request, redirect, url_for, flash, session
 from flask_mysqldb import MySQL
 
-from __main__ import app, mysql, e_id
+from __main__ import app, mysql
+
+def verificar(id):
+    if 'loggedin' not in session:
+        return False
+    if session['type'] == "encuestado":
+        return False
+    cur = mysql.connection.cursor()
+    cur.execute("SELECT * FROM encuesta WHERE id_encuesta = %s AND id_encuestador = %s;",(id, session['id'],))
+    validar = cur.fetchall()
+    if validar == ():
+        return False
+    return True
 
 @app.route('/edit_respuesta/<id>/<id_pregunta>/<id_respuesta>', methods=['GET','POST'])
 def edit_respuesta(id,id_pregunta,id_respuesta):
+    # verificar
+    if not verificar(id):
+        flash("Usted no puede editar esta encuesta.")
+        return redirect(url_for("forms"))
+    # editar respuesta
     if request.method == 'POST':
         if(request.form['respuesta']!=""):
             cur = mysql.connection.cursor()
@@ -20,6 +36,11 @@ def edit_respuesta(id,id_pregunta,id_respuesta):
 
 @app.route('/edit_pregunta/<id>/<id_pregunta>', methods=['GET','POST'])
 def edit_pregunta(id,id_pregunta):
+    # verificar
+    if not verificar(id):
+        flash("Usted no puede editar esta encuesta.")
+        return redirect(url_for("forms"))
+    #editar pregunta
     if request.method == 'POST':
         if(request.form['pregunta']!=""):
             cur = mysql.connection.cursor()
@@ -34,6 +55,11 @@ def edit_pregunta(id,id_pregunta):
 
 @app.route('/edit_titulo/<id>', methods=['GET','POST'])
 def edit_titulo(id):
+    # verificar
+    if not verificar(id):
+        flash("Usted no puede editar esta encuesta.")
+        return redirect(url_for("forms"))
+    # editar titulo
     if request.method == 'POST':
         if(request.form['titulo']!=""):
             cur = mysql.connection.cursor()
@@ -49,9 +75,14 @@ def edit_titulo(id):
 
 @app.route('/new_alternativa/<id>/<id_pregunta>', methods=['GET','POST'])
 def new_alternativa(id,id_pregunta):
+    # verificar
+    if not verificar(id):
+        flash("Usted no puede editar esta encuesta.")
+        return redirect(url_for("forms"))
+
+    #agregar alternativa
     if request.method == 'POST':
         if(request.form['alternativa']!=""):
-            print(request.form['alternativa'])
             cur = mysql.connection.cursor()
             cur.execute("INSERT INTO respuesta (id_pregunta, texto_respuesta) VALUES (%s , %s);", (id_pregunta ,request.form['alternativa'],))
             mysql.connection.commit()
@@ -63,9 +94,13 @@ def new_alternativa(id,id_pregunta):
 
 @app.route('/delete_alternativa/<id>/<id_alternativa>')
 def delete_alternativa(id,id_alternativa):
-    cur = mysql.connection.cursor()
-    
+    # verificar
+    if not verificar(id):
+        flash("Usted no puede editar esta encuesta.")
+        return redirect(url_for("forms"))
+
     #borrar la respuesta
+    cur = mysql.connection.cursor()
     cur.execute("DELETE FROM respuesta WHERE id_respuesta = %s;", (id_alternativa,))
     mysql.connection.commit()
     flash("Respuesta eliminada.")
@@ -73,9 +108,13 @@ def delete_alternativa(id,id_alternativa):
 
 @app.route('/delete_pregunta/<id>/<id_pregunta>')
 def delete_pregunta(id,id_pregunta):
-    cur = mysql.connection.cursor()
+    # verificar
+    if not verificar(id):
+        flash("Usted no puede editar esta encuesta.")
+        return redirect(url_for("forms"))
 
     #borrar todas las respuestas vinculados a la pregunta
+    cur = mysql.connection.cursor()
     cur.execute("DELETE FROM respuesta WHERE id_pregunta = %s;", (id_pregunta,))
     mysql.connection.commit()
     
@@ -87,7 +126,7 @@ def delete_pregunta(id,id_pregunta):
 
 @app.route('/edit/<id>', methods=['GET','POST'])
 def edit(id):
-
+    cur = mysql.connection.cursor()
     #Prohibido acceder a personas que no están registradas.
     if 'loggedin' not in session:
         return(render_template("403.html"))
@@ -97,17 +136,13 @@ def edit(id):
     #Los encuestados tampoco deberían editar las encuestas.
     if(tiposesion == "encuestado"):
         return(render_template("403.html"))
+    
+    # verificar
+    if not verificar(id):
+        flash("Usted no puede editar esta encuesta.")
+        return redirect(url_for("forms"))
 
-    #Si la encuesta esta cerrada, no se debe poder editar.
-    cur = mysql.connection.cursor()
-
-    cur.execute("SELECT cerrada FROM encuesta WHERE id_encuesta = " + id)
-    cerrada = cur.fetchall()
-    mysql.connection.commit()
-
-    if(cerrada[0][0] == 1):
-        return(render_template("403.html"))
-
+    #edit
     if request.method == 'POST':
         if(request.form['pregunta'] != ""):
             tipo_pregunta = -1;
@@ -127,8 +162,6 @@ def edit(id):
 
     #Visualización
     else:
-        cur = mysql.connection.cursor()
-
         #Encontrar el nombre de la encuesta en db
         cur.execute("SELECT titulo FROM encuesta WHERE id_encuesta = " + id)
         nombre_encuesta = cur.fetchall()
@@ -152,12 +185,15 @@ def edit(id):
             #esto recibiría el id_pregunta, id_encuesta, respuestas, tipo y el texto.
             preguntas.append([pregunta[0], pregunta[1], respuestas, pregunta[2], pregunta[4]])
         mysql.connection.commit()
-        
         return render_template('/encuestadores/edit.html', form = preguntas, titulo = nombre_encuesta[0][0], id = id)
 
 @app.route('/cerrar_encuesta/<id>', methods=['GET','POST'])
 def cerrar_encuesta(id):
-    
+    # verificar
+    if not verificar(id):
+        flash("Usted no puede Cerrar esta encuesta.")
+        return redirect(url_for("forms"))
+
     cur = mysql.connection.cursor()
     cur.execute("UPDATE encuesta SET cerrada = '1' WHERE encuesta.id_encuesta = " + id)
